requirements.txt
# Core LLM APIs
anthropic>=0.40.0
openai>=1.0.0
requests>=2.31.0       # Ollama HTTP API

# Dataset
datasets>=2.14.0

# NLI model
torch>=2.1.0
transformers>=4.36.0
accelerate>=0.24.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.66.0
rich>=13.0.0
PyYAML>=6.0

# Testing
pytest>=8.0.0

# Evaluation & visualization
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.13.0
